remove suite hook configuration and make information flow one-way
