+++ pkg/test_suite_definition/cni.go
@@ -1,4 +1,4 @@
-package main
+package test_suite_definition
 
 import (
 	"strings"
