+++ pkg/test_suite_definition/e2e_test.go
@@ -1,4 +1,4 @@
-package main
+package test_suite_definition
 
 import (
 	"fmt"
